Refactor image modal into reusable ImageViewerModal component
